.history/requirements_20231027165507.txt
fastapi
uvicorn[standard]
motor
sendgrid
python-jose[cryptography]
python-multipart
passlib
jinja2
pillow
bcrypt
pydantic[email]
joe